+++ client/src/components/Auth/RegistrationForm/Steps/stepLabels.js
@@ -1,0 +1,3 @@
+const stepLabels = ["Email and username", "Password", "Confirmation"];
+
+export default stepLabels;
